move all routes to Router component
